requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0          # Async MongoDB driver
pymongo==4.7.2
python-dotenv==1.0.1
anthropic==0.25.3
pydantic==2.7.1
httpx==0.27.0         # Required by FastAPI test client
pytest==8.2.0
pytest-asyncio==0.23.6
